update to big decimal

+++ LoadData/src/main/java/gupiao/china/LoadHistoricalDataToDB.java
@@ -110,8 +110,8 @@ public class LoadHistoricalDataToDB {
 				psL.setFloat(5, stockDeal.getClosePrice());
 				psL.setFloat(6, stockDeal.getHighestPrice());
 				psL.setFloat(7, stockDeal.getLowestPrice());
-				psL.setFloat(8, stockDeal.getDealAmount());
-				psL.setLong(9, stockDeal.getDealNumber());
+				psL.setBigDecimal(8, stockDeal.getDealAmount());
+				psL.setBigDecimal(9, stockDeal.getDealNumber());
 	
 				psL.execute();
 				conn.commit();
